backend/requirements.txt
# Zenkyra Backend V1
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

pydantic==2.10.4
pydantic-settings==2.7.0

SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

redis==5.2.1

qdrant-client==1.12.1

google-generativeai==0.8.3

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

tenacity==9.0.0
httpx==0.28.1

pypdf==5.1.0
email-validator==2.2.0
